Move Function: combineOfficeRnDData()
